chore(sc-cli): improve runner and signals
